# api/openapi.yaml
openapi: 3.1.0
info:
  title: StealthBridge Testnet API
  version: 0.3.0
  description: >
    Explicitly read-only public Stellar RPC observation and operator-configured
    corridor discovery. Confidential settlement submission is disabled.
    No hardcoded quotes, corridors, payment providers or successful payouts.
servers:
  - url: http://localhost:8080
paths:
  /health:
    get:
      operationId: health
      summary: API process liveness; not chain readiness
      responses:
        "200": {description: API process responding}
  /ready:
    get:
      operationId: readiness
      summary: Network/database readiness; not payment eligibility
      responses:
        "200": {description: Both external read dependencies available}
        "503": {description: One or more dependencies unavailable}
  /v1/observer:
    get:
      operationId: observerHead
      summary: Last persisted, opt-in Stellar Testnet ledger checkpoint
      responses:
        "200":
          description: Last observed public ledger metadata (not necessarily fresh)
          content:
            application/json:
              schema: {$ref: "#/components/schemas/LedgerCheckpoint"}
        "404": {description: No observer has recorded a verified ledger yet}
        "503": {description: Database not configured or unavailable}
  /v1/network:
    get:
      operationId: network
      summary: Read observed Stellar Testnet ledger after verifying network passphrase
      responses:
        "200":
          description: Observed ledger metadata
          content:
            application/json:
              schema: {$ref: "#/components/schemas/NetworkStatus"}
        "502": {description: RPC unreachable, malformed or wrong network}
  /v1/capabilities:
    get:
      operationId: capabilities
      summary: Truthful capability flags, never inferred from RPC liveness
      responses:
        "200":
          description: Current integration flags
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Capabilities"}
  /v1/corridors:
    get:
      operationId: corridors
      summary: List enabled, operator-created corridor entries, potentially empty
      responses:
        "200":
          description: Registered enabled corridors
          content:
            application/json:
              schema:
                type: array
                items: {$ref: "#/components/schemas/Corridor"}
        "503": {description: Database not configured or temporarily unavailable}
  /v1/corridors/page:
    get:
      operationId: corridorPage
      summary: Read stable keyset pages of enabled, operator-configured corridors
      parameters:
        - name: after
          in: query
          description: Previous next_cursor UUID; omit on first page
          schema: {type: string, format: uuid}
        - name: limit
          in: query
          description: Items per page, default 25, maximum 100
          schema: {type: integer, minimum: 1, maximum: 100, default: 25}
      responses:
        "200":
          description: Page from current configured database state; no FX or payout claims
          content:
            application/json:
              schema: {$ref: "#/components/schemas/CorridorPage"}
        "400": {description: Invalid limit or cursor}
        "503": {description: Database not configured or unavailable}
  /v1/corridors/{id}:
    get:
      operationId: corridorById
      summary: Retrieve a single enabled, operator-configured corridor
      parameters:
        - name: id
          in: path
          required: true
          schema: {type: string, format: uuid}
      responses:
        "200":
          description: Configured enabled corridor, not proof of real partner/liquidity eligibility
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Corridor"}
        "400": {description: Invalid UUID}
        "404": {description: Unknown or disabled corridor}
        "503": {description: Database missing or unavailable}
  /v1/transactions/{hash}:
    get:
      operationId: transactionObservation
      summary: Read public Testnet transaction status without exposing raw XDR or events
      parameters:
        - name: hash
          in: path
          required: true
          description: Exact 64-character hex transaction hash
          schema: {type: string, pattern: "^[a-fA-F0-9]{64}$"}
      responses:
        "200":
          description: Transaction included in retained Stellar RPC history (success or failed execution)
          content:
            application/json:
              schema: {$ref: "#/components/schemas/TransactionObservation"}
        "400": {description: Invalid hash}
        "404": {description: Hash not found within RPC retention; not proof it never existed}
        "502": {description: Upstream RPC failure or unexpected response}
  /v1/settlements:
    post:
      operationId: disabledSettlement
      summary: Reserved future authentication-bound settlement command, disabled
      responses:
        "501": {description: Fund-moving functionality not enabled}
components:
  schemas:
    Capabilities:
      type: object
      required: [payments_enabled, confidential_token_verified, private_payments_verified, fiat_payouts_enabled]
      properties:
        payments_enabled: {type: boolean}
        confidential_token_verified: {type: boolean}
        private_payments_verified: {type: boolean}
        fiat_payouts_enabled: {type: boolean}
    LedgerCheckpoint:
      type: object
      required: [ledger_sequence, ledger_hash, ledger_closed_at_unix, source]
      properties:
        ledger_sequence: {type: integer, minimum: 1}
        ledger_hash: {type: string, pattern: "^[a-f0-9]{64}$"}
        ledger_closed_at_unix: {type: string}
        source: {type: string, const: stellar-rpc}
    NetworkStatus:
      type: object
      required: [network, passphrase, protocol_version, ledger_sequence, ledger_closed_at_unix, ledger_hash, source]
      properties:
        network: {type: string, const: testnet}
        passphrase: {type: string}
        protocol_version: {type: integer}
        ledger_sequence: {type: integer}
        ledger_closed_at_unix: {type: string}
        ledger_hash: {type: string}
        source: {type: string, const: stellar-rpc}
    Corridor:
      type: object
      required: [id, origin_country, destination_country, asset_code, privacy_rail]
      properties:
        id: {type: string, format: uuid}
        origin_country: {type: string}
        destination_country: {type: string}
        asset_code: {type: string}
        asset_issuer: {type: [string, "null"]}
        privacy_rail: {type: string, enum: [confidential-token, private-payments]}
    CorridorPage:
      type: object
      required: [items, next_cursor]
      properties:
        items:
          type: array
          maxItems: 100
          items: {$ref: "#/components/schemas/Corridor"}
        next_cursor: {type: [string, "null"], format: uuid}
    TransactionObservation:
      type: object
      required: [hash, status, ledger, closed_at_unix, latest_ledger, source]
      properties:
        hash: {type: string, pattern: "^[0-9a-f]{64}$"}
        status: {type: string, enum: [SUCCESS, FAILED]}
        ledger: {type: integer}
        closed_at_unix: {type: string}
        latest_ledger: {type: integer}
        source: {type: string, const: stellar-rpc}